BUGFIX: broken merge builds
When GitlabWebHook::Project objects are instantiated from a publisher, the
matching_scms array is never filled, and operations related to the configuration
of the git plugin must be modified.
The simple approach used will make not very confident the behaviour of gitlab
notifier when running on multiple-scms project, specially if merged branches
are configured.